lesson-5

+++ src/app/common/component/modal/modal.component.ts
@@ -46,7 +46,7 @@ export class ModalComponent implements OnInit {
     if (code !== 27) {
       return;
     }
-    this._modalContext.destroy();
+    this._modalContext && this._modalContext.destroy();
     this.isOpen = false;
 
   }
